+++ Riskeer/Common/test/Riskeer.Common.Forms.Test/Views/WaveHeightCalculationsViewTest.cs
@@ -264,7 +264,7 @@ namespace Riskeer.Common.Forms.Test.Views
             var guiService = Substitute.For<IHydraulicBoundaryLocationCalculationGuiService>();
 
             HydraulicBoundaryLocationCalculation[] performedCalculations = null;
-            guiService.When(_ => _.CalculateWaveHeights(Arg.Any<IEnumerable<HydraulicBoundaryLocationCalculation>>(), Arg.Any<IAssessmentSection>(), Arg.Any<int>(), Arg.Any<string>()))
+            guiService.When(_ => _.CalculateWaveHeights(Arg.Any<IEnumerable<HydraulicBoundaryLocationCalculation>>(), Arg.Any<IAssessmentSection>(), Arg.Any<double>(), Arg.Any<string>()))
                       .Do(callInfo =>
                       {
                           performedCalculations = callInfo.Arg<IEnumerable<HydraulicBoundaryLocationCalculation>>().ToArray();
@@ -278,7 +278,7 @@ namespace Riskeer.Common.Forms.Test.Views
             // Assert
             Assert.AreEqual(1, performedCalculations.Length);
             Assert.AreSame(hydraulicBoundaryLocationCalculations.First(), performedCalculations.First());
-            guiService.Received().CalculateWaveHeights(Arg.Any<IEnumerable<HydraulicBoundaryLocationCalculation>>(), Arg.Any<IAssessmentSection>(), Arg.Any<int>(), Arg.Any<string>());
+            guiService.Received().CalculateWaveHeights(Arg.Any<IEnumerable<HydraulicBoundaryLocationCalculation>>(), Arg.Any<IAssessmentSection>(), Arg.Any<double>(), Arg.Any<string>());
         }
 
         [Test]
@@ -316,7 +316,7 @@ namespace Riskeer.Common.Forms.Test.Views
             HydraulicBoundaryLocationCalculation[] performedCalculations = null;
             double targetProbabilityValue = double.NaN;
             string calculationIdentifierValue = null;
-            guiService.When(_ => _.CalculateWaveHeights(Arg.Any<IEnumerable<HydraulicBoundaryLocationCalculation>>(), Arg.Any<IAssessmentSection>(), Arg.Any<int>(), Arg.Any<string>()))
+            guiService.When(_ => _.CalculateWaveHeights(Arg.Any<IEnumerable<HydraulicBoundaryLocationCalculation>>(), Arg.Any<IAssessmentSection>(), Arg.Any<double>(), Arg.Any<string>()))
                       .Do(callInfo =>
                       {
                           performedCalculations = callInfo.Arg<IEnumerable<HydraulicBoundaryLocationCalculation>>().ToArray();
